app/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
pyserial==3.5
python-dateutil==2.9.0
gunicorn==22.0.0
